frontend/swdc/requirements.txt
pip
wheel
django-hamlpy
libsass
